bruk av flere komponenter
