+++ src/minoversikt/motekalender.less
@@ -2,6 +2,3 @@
   text-align: center;
   margin-bottom: 0.2rem;
 }
-
-.motekalender_popover {
-}
