+++ src/utils/dato-utils.ts
@@ -169,3 +169,25 @@ export function oppfolgingStartetDato(dato: string): Maybe<Date> {
 
     return oppfolgingStartetDato < tidligsteDato ? tidligsteDato : oppfolgingStartetDato;
 }
+
+export function dagFraDato(dato: Date): string {
+    const i = dato.getDay();
+    switch (i) {
+        case 0:
+            return 'Søndag';
+        case 1:
+            return 'Mandag';
+        case 2:
+            return 'Tirsdag';
+        case 3:
+            return 'Onsdag';
+        case 4:
+            return 'Torsdag';
+        case 5:
+            return 'Fredag';
+        case 6:
+            return 'Lørdag';
+        default:
+            return '...';
+    }
+}
